feature/ATC-18 - moves .choose() and .choose_weight() to new generalUtilities file and updates references
